TLS parameters have been moved to a subcommand of the deployment command.
